:recycle: Move 'applyPagination' to Services
